+++ apps/blackjack-client/src/main/java/UI/PH_Gameplay/Model/TrumpChip.java
@@ -1,4 +1,4 @@
-package UI.Demo.Model;
+package UI.PH_Gameplay.Model;
 
 import javax.swing.ImageIcon;
